requirements.txt
Django>=3.2.0
djangorestframework>=3.2,<4.0
django-redis>=5.0,<6.0
uWSGI>=2.0.19.1,<2.1
djangorestframework-simplejwt
uWSGI>=2.0.19.1,<2.1
rest_framework_simplejwt
celery
